+++ ariac_db/CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 2025.1.2 (2025-09-26)
 ----------------------
 * Update scoring_weights.yaml
+* Set bonuses to zero if run was aborted (ENDED without completing all orders)
 
 2025.1.1 (2025-09-19)
 ----------------------